// pslpwcl/src/app/modules/pslpwcl-common/models/type-application-card.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
export enum TypeApplicationCard {
  Home = 'H',
  Cittadino = 'C',
  Fascicolo = 'F',
  BackOffice = 'B'
}