// pslpwcl/src/app/modules/pslpwcl-common/models/userInfo.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * Portale Servizi Lavoro API
 * API per il backend del portale servizi lavoro.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface UserInfo { 
    /**
     * nome dell'utente
     */
    nome?: string;
    /**
     * cognome dell'utente
     */
    cognome?: string;
    /**
     * codice fiscale dell'utente
     */
    codFisc?: string;
    /**
     * nome dell'ente
     */
    ente?: string;
    /**
     * nome del ruolo dell'utente
     */
    ruolo?: string;
    /**
     * codice del ruolo dell'utente
     */
    cod_ruolo?: string;
    /**
     * livello di autenticazione utilizzato per accedere
     */
    livAuth?: number;
    /**
     * community di utenti utilizzata per accedere
     */
    community?: string;
}